docker/script_docker/requirements.txt
flask==1.1.1
pandas
mysql-connector-python